// source/frontend/src/models/api/generated/ApiGen_CodeTypes_LeaseStakeholderTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_LeaseStakeholderTypes {
  ASGN = 'ASGN',
  OWNER = 'OWNER',
  OWNREP = 'OWNREP',
  PMGR = 'PMGR',
  REP = 'REP',
  TEN = 'TEN',
  UNK = 'UNK',
}