// source/frontend/src/models/api/generated/ApiGen_Concepts_CompReqLeaseStakeholder.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_LeaseStakeholder } from './ApiGen_Concepts_LeaseStakeholder';

// LINK: @backend/apimodels/Models/Concepts/CompensationRequisition/CompReqLeaseStakeholderModel.cs
export interface ApiGen_Concepts_CompReqLeaseStakeholder extends ApiGen_Base_BaseAudit {
  compReqLeaseStakeholderId: number | null;
  compensationRequisitionId: number;
  leaseStakeholderId: number;
  leaseStakeholder: ApiGen_Concepts_LeaseStakeholder | null;
}